Feat/keumbi2

+++ server/src/main/java/com/pre007/server/answer/service/AnswerService.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -46,10 +47,13 @@ public class AnswerService {
     public Answer updateAnswer(Answer answer) {
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
 
+        Member postMember = memberService.findVerifiedMember(findAnswer.getMember().getMemberId()); // 작성자
+        if(!Objects.equals(memberService.getLoginMember().getMemberId(), postMember.getMemberId())) // 로그인 유저 != 작성자
+            throw new BusinessLogicException(ExceptionCode.MEMBER_UNAUTHORIZED); // 수정 권한 없음
         Optional.ofNullable(answer.getAnswerContent())
                 .ifPresent(answerContent -> findAnswer.setAnswerContent(answerContent));
-        Optional.ofNullable(answer.getAnswerCheck())
-                .ifPresent(answerCheck -> findAnswer.setAnswerCheck(answerCheck));
+       /* Optional.ofNullable(answer.getAnswerCheck())
+                .ifPresent(answerCheck -> findAnswer.setAnswerCheck(answerCheck));*/
         //findAnswer.setModifiedAt(LocalDateTime.now());
         Answer savedAnswer = answerRepository.save(findAnswer);
 
@@ -70,6 +74,11 @@ public class AnswerService {
     }
     public void deleteOneAnswer(long answerId) {
         Answer findAnswer = findVerifiedAnswer(answerId);
+
+        Member postMember = memberService.findVerifiedMember(findAnswer.getMember().getMemberId()); // 작성자
+        if(!Objects.equals(memberService.getLoginMember().getMemberId(), postMember.getMemberId())) // 로그인 유저 != 작성자
+            throw new BusinessLogicException(ExceptionCode.MEMBER_UNAUTHORIZED); // 수정 권한 없음
+
         answerRepository.delete(findAnswer);
     }
     @Transactional(readOnly = true)
